move conf to test-config, helpers to project root
