+++ karma.conf.js
@@ -1,2 +1,2 @@
 // Look in ./config for karma.conf.js
-module.exports = require('./config/karma.conf.js');
+module.exports = require('./test-config/karma.conf.js');
