+++ protractor.conf.js
@@ -1,2 +1,2 @@
 // look in ./config for protractor.conf.js
-exports.config = require('./config/protractor.conf.js').config;
+exports.config = require('./test-config/protractor.conf.js').config;
