+++ webpack.config.ts
@@ -29,7 +29,7 @@ const webpackMerge = require('webpack-merge');
 
 const includeClientPackages = require('./config/helpers.js').includeClientPackages;
 const hasProcessFlag = require('./config/helpers.js').hasProcessFlag;
-const root = require('./config/helpers.js').root;
+const root = require('./helpers.js').root;
 
 const ENV = process.env.npm_lifecycle_event;
 const AOT = ENV === 'build:aot' || ENV === 'build:aot:dev' || ENV === 'server:aot' || ENV === 'watch:aot';
